requirements.txt
django==4.2.4
psycopg2-binary==2.9.6
tabula-py==2.7.0
django-htmx==1.16.0
django-admin-interface==0.26.0